re-organized commit 3, migration of dev env
